pyproject.toml
[project]
name = "13-advanced-retrieval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-experimental>=0.3.4",
    "langchain>=0.3.19",
    "langchain-cohere==0.4.4",
    "cohere>=5.12.0,<5.13.0", 
    "langchain-openai>=0.3.7",
    "qdrant-client>=1.13.2",
    "rank-bm25>=0.2.2",
    "langchain-qdrant>=0.2.0",
    "ragas==0.3.6",
    "rapidfuzz>=3.0.0",
    "loguru>0.7.0",
    "matplotlib>=3.10.0",
    "jinja2>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
